requirements.txt
python-chess
openai
pyttsx3
cryptography
black